% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend_title.R
\name{add_legend_title}
\alias{add_legend_title}
\title{Adds legend title to first cell in README sheet}
\usage{
add_legend_title(wb, title)
}
\arguments{
\item{wb}{Workbook object from openxlsx package}

\item{title}{Character string with title from spreadsheet()}
}
\description{
Adds legend title to first cell in README sheet
}
