% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_main_sheets.R
\name{add_main_sheets}
\alias{add_main_sheets}
\title{Add main sheets to spreadsheet}
\usage{
add_main_sheets(wb, spreadsheet)
}
\arguments{
\item{wb}{openxlsx workbook object}

\item{spreadsheet}{spreadsheet object from spreadsheet()}
}
\description{
Add main sheets to spreadsheet
}
