% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_meta}
\alias{add_meta}
\title{Function to add column description metadata to README file}
\usage{
add_meta(wb, spreadsheet, nextFreeRow)
}
\arguments{
\item{wb}{openxlsx workbook object}

\item{spreadsheet}{spreadsheet object returned from spreadsheet()}

\item{nextFreeRow}{index of row with empty cell after writing sheet legends}
}
\description{
Function to add column description metadata to README file
}
