% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_readme.R
\name{add_readme}
\alias{add_readme}
\title{Creates README sheet in first sheet}
\usage{
add_readme(wb)
}
\arguments{
\item{wb}{Workbook object from openxlsx package}
}
\description{
Creates README sheet in first sheet
}
\examples{
\dontrun{
add_readme(wb)
}
}
