% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet_legend.R
\name{add_sheet_legend}
\alias{add_sheet_legend}
\title{Add legends for sheets below main legend}
\usage{
add_sheet_legend(wb, spreadsheet)
}
\arguments{
\item{wb}{openxlsx workbook object}

\item{spreadsheet}{spreadsheet object returned from spreadsheet()}
}
\value{
numeric value with row index for next empty cell
}
\description{
Add legends for sheets below main legend
}
