% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_meta.R
\name{append_meta}
\alias{append_meta}
\title{Annotate Data Frame with Column Descriptions}
\usage{
append_meta(results, colname_descriptions)
}
\arguments{
\item{results}{A data.frame to be marked up with comments for each column.
generated.}

\item{colname_descriptions}{A named character vector where names match column
names of \code{results} and values are their descriptions.}
}
\value{
The input data.frame with comments added to each column.
}
\description{
Adds descriptions as comments to all columns in a given data frame.
}
\examples{

results <- append_meta(
  results = iris,
  colname_descriptions = c(
    "Sepal.Length" = "Length of the sepal in cm",
    "Sepal.Width" = "Width of the sepal in cm",
    "Petal.Length" = "Length of the petal in cm",
    "Petal.Width" = "Width of the petal in cm",
    "Species" = "Species of iris"
  )
)

# Example to show column descriptions are stored as comments
comment(results$Sepal.Length)

}
