% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta.R
\name{create_meta}
\alias{create_meta}
\title{Create .cols sidecar meta data file}
\usage{
create_meta(file_name, table_variable_name, colname_descriptions)
}
\arguments{
\item{file_name}{A character string specifying the path and file name (including the .csv part) of the corresponding table meta data is created for.}

\item{table_variable_name}{A data frame for which metadata is being generated.}

\item{colname_descriptions}{A named character vector where names correspond to column names of \code{table_variable_name} and values are their descriptions.}
}
\value{
Writes a \code{.cols} metadata file as a side effect. No object is returned.
}
\description{
This function creates a .cols sidecar meta data file of column name descriptions for a given table.
It checks if all column names in the table variable name are described in the colname_descriptions and stops with an error message if not.
The .cols file is saved in the same directory as the table variable name with the same name as the file name.
}
