% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelerate.R
\name{excelerate}
\alias{excelerate}
\title{Create and save supplementary tables as excel files}
\usage{
excelerate(...)
}
\arguments{
\item{...}{One or more spreadsheet objects returned from the \code{spreadsheet()} function.}
}
\description{
Create and save supplementary tables as excel files
}
