% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_path.R
\name{get_file_path}
\alias{get_file_path}
\title{Get all the files in a directory from the path that match a pattern}
\usage{
get_file_path(path, pattern)
}
\arguments{
\item{path}{path to directory where files for supplementary tables exist}

\item{pattern}{pattern to subset files in the directory}
}
\value{
Character vector of full paths to files
}
\description{
Get all the files in a directory from the path that match a pattern
}
\examples{
\dontrun{
get_file_path("path/to/data/", "extract_matching_files")
}
}
