% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results}
\alias{read_results}
\title{Function to read in the results tables and sidecar .cols files}
\usage{
read_results(full_path, sheet_legend)
}
\arguments{
\item{full_path}{character string containing full path to csv or tsv}

\item{sheet_legend}{character string containing table legend}
}
\value{
List of dataframes containing the main results (data frame),
the metadata (data frame) and sheet legend (character string)
}
\description{
Function to read in the results tables and sidecar .cols files
}
\examples{
\dontrun{
read_results("path/to/data/", "table legend")
}
}
