% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_meta.R
\name{append_meta}
\alias{append_meta}
\title{Annotate Data Frame with Column Descriptions}
\usage{
append_meta(results, colname_descriptions)
}
\arguments{
\item{results}{A data.frame to be marked up with comments for each column.
generated.}

\item{colname_descriptions}{A named character vector where names match column
names of \code{results} and values are their descriptions.}
}
\value{
The input data.frame with comments added to each column.
}
\description{
Adds descriptions as comments to all columns in a given data frame.
}
\examples{

append_meta(
  results = mtcars,
  colname_descriptions = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)

}
