% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta.R
\name{create_meta}
\alias{create_meta}
\title{Create .cols Sidecar Metadata File}
\usage{
create_meta(results, colname_descriptions)
}
\arguments{
\item{results}{A data.frame to be marked up with comments for each column.
generated.}

\item{colname_descriptions}{A named character vector where names match column
names of \code{results} and values are their descriptions.}
}
\value{
A data.frame marked up with comments for each column.
}
\description{
This function creates a .cols sidecar metadata file with column name
descriptions for a given table. It checks if all column names in the table
are described. The \code{.cols} file is saved in the same directory as the data
file, appending \code{.cols} to the original file name (e.g.,
\code{your_file.csv.cols}).
}
\examples{

create_meta(
  results = mtcars,
  colname_descriptions = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)

}
