% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelerate-package.R
\docType{package}
\name{excelerate-package}
\alias{excelerate-package}
\title{excelerate: Functions to Create and Export Excel Spreadsheets for Supplementary Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The **excellerate** package provides a set of functions to easily create and export Excel spreadsheets, specifically for generating supplementary tables required when preparing research manuscripts for scientific journals. It allows for quick and customizable table creation from R data, making it easier to structure and format tables according to journal requirements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ameliaes.github.io/excelerate/}
}

}
\author{
\strong{Maintainer}: Amelia Edmondson-Stait \email{aedmond3@ed.ac.uk} (\href{https://orcid.org/0000-0003-1456-5128}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mark James Adams \email{mark.adams@ed.ac.uk} (\href{https://orcid.org/0000-0002-3599-6018}{ORCID})
}

}
\keyword{internal}
