% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelerate-package.R
\docType{package}
\name{excelerate-package}
\alias{excelerate-package}
\title{excelerate: Create Excel Spreadsheets for Supplementary Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The excelerate package simplifies the creation of supplementary tables for research papers. It uses the openxlsx R package to produce tables in a consistent format. You can easily turn data frames into Excel files, with accompanying metadata in a README sheet, and save them as numbered supplementary files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ameliaes.github.io/excelerate/}
}

}
\author{
\strong{Maintainer}: Amelia Edmondson-Stait \email{aedmond3@ed.ac.uk} (\href{https://orcid.org/0000-0003-1456-5128}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mark James Adams \email{mark.adams@ed.ac.uk} (\href{https://orcid.org/0000-0002-3599-6018}{ORCID})
}

}
\keyword{internal}
