% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelerate.R
\name{excelerate}
\alias{excelerate}
\title{Create and Save Excel Files with Supplementary Tables}
\usage{
excelerate(...)
}
\arguments{
\item{...}{One or more \code{spreadsheet} objects created with \code{\link[=spreadsheet]{spreadsheet()}}.}
}
\description{
This function creates and saves Excel files from given spreadsheet objects.
}
\examples{
temp_dir <- tempdir()
temp_file <- file.path(temp_dir, "example.csv")
write.csv(mtcars, temp_file, row.names = FALSE)
create_meta(
  file_name = temp_file,
  table_variable_name = mtcars,
  colname_descriptions = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)
sheet1 <- sheet(temp_dir, "example\\\\.csv$", "Sheet A", "Legend A")
sheet2 <- sheet(temp_dir, "example\\\\.csv$", "Sheet B", "Legend B")
sp <- spreadsheet(
  "Supplementary Table X",
  file.path(temp_dir, "example.xlsx"),
  sheet1, sheet2)
excelerate(sp)

# Clean up the temporary files
unlink(c(temp_file,
  paste0(temp_file, ".cols"),
  paste0(temp_dir, "/example.xlsx")))
}
