% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelerate.R
\name{excelerate}
\alias{excelerate}
\title{Create and Save Excel Files with Supplementary Tables}
\usage{
excelerate(...)
}
\arguments{
\item{...}{One or more \code{spreadsheet} objects created with \code{\link[=spreadsheet]{spreadsheet()}}.}
}
\description{
This function creates and saves Excel files from given spreadsheet objects.
}
\examples{
temp_dir <- tempdir()

results <- append_meta(
  results = mtcars,
  colname_descriptions = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)

sheet1 <- sheet(results, "Sheet A", "Legend A")
sheet2 <- sheet(results, "Sheet B", "Legend B")

sp <- spreadsheet(
  "Supplementary Table X",
  file.path(temp_dir, "example.xlsx"),
  sheet1, sheet2)

excelerate(sp)

# Clean up the temporary files
unlink(paste0(temp_dir, "/example.xlsx"))
}
