% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelerate.R
\name{excelerate}
\alias{excelerate}
\title{Create and Save Excel Files with Supplementary Tables}
\usage{
excelerate(...)
}
\arguments{
\item{...}{One or more \code{spreadsheet} objects created with \code{\link[=spreadsheet]{spreadsheet()}}.}
}
\description{
This function creates and saves Excel files from given spreadsheet objects.
}
\examples{
temp_dir <- tempdir()

results <- append_meta(
  results = iris,
  colname_descriptions = c(
    "Sepal.Length" = "Length of the sepal in cm",
    "Sepal.Width" = "Width of the sepal in cm",
    "Petal.Length" = "Length of the petal in cm",
    "Petal.Width" = "Width of the petal in cm",
    "Species" = "Species of iris"
  )
)

sheet1 <- sheet(results, "Sheet A", "Legend A")
sheet2 <- sheet(results, "Sheet B", "Legend B")

sp <- spreadsheet(
  "Supplementary Table X",
  file.path(temp_dir, "example.xlsx"),
  sheet1, sheet2
)

excelerate(sp)

# Clean up the temporary files
unlink(file.path(temp_dir, "example.xlsx"))
}
