% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.R
\name{sheet}
\alias{sheet}
\title{Read in file path, file pattern, sheet name and sheet legend}
\usage{
sheet(path, pattern, sheet_name, sheet_legend)
}
\arguments{
\item{path}{character string with path to directory where files for supplementary tables exist}

\item{pattern}{character string with pattern to identify one file in the directory}

\item{sheet_name}{character string sheet name for table}

\item{sheet_legend}{character string containing table legend}
}
\value{
Named nested list of dataframes containing the main results (data frame),
the metadata (data frame) and sheet legend (character string).
Named by sheet_name.
}
\description{
Read in file path, file pattern, sheet name and sheet legend
}
\examples{
\dontrun{
sheet("path/to/data/", "file_name", "sheet name", "sheet legend")
}
}
