% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.R
\name{sheet}
\alias{sheet}
\title{Append column descriptions as comments to data frame}
\usage{
sheet(results, sheet_name, sheet_legend)
}
\arguments{
\item{results}{A data.frame marked up with comments for each column.}

\item{sheet_name}{A character string for naming the sheet.}

\item{sheet_legend}{A character string containing the table legend.}
}
\value{
A named nested list containing the main results (data frame),
metadata (data frame), and sheet legend. Named by \code{sheet_name}.
}
\description{
This function annotates your data frame with descriptions for each column,
which is then inserted as a table in the first sheet ("README" sheet)
of your Excel spreadsheet.
}
\examples{

results <- append_meta(
  results = mtcars,
  colname_descriptions = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)
sheet(results, "Example Sheet", "An example legend")

}
