% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadheet.R
\name{spreadsheet}
\alias{spreadsheet}
\title{Function to combine sheets into a list, with supplementary table legend and
filename to save excel.}
\usage{
spreadsheet(..., title, filename)
}
\arguments{
\item{...}{A list of sheets}

\item{title}{Character string with legend title for spreadsheet}

\item{filename}{Character string with path and file name for excel
spreadsheet}
}
\value{
List of sheets, with title and path for saving excel file
}
\description{
Function to combine sheets into a list, with supplementary table legend and
filename to save excel.
}
\examples{
\dontrun{
spreadsheet(
sheet("path/to/data/",
      "file_name_for_first_sheet",
      "sheet name A",
      "Table legend A"),
sheet("path/to/data/second_file",
      "file_name_for_second_sheet",
      "sheet name B",
      "Table legend B"),
"Supplementary Table Legend Title",
"path/to/save/excel/file_name")
}
}
